requirements.txt
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Mail==0.10.0
Flask-WTF==1.2.1
PyMySQL==1.1.2
cryptography==46.0.5
python-dotenv==1.2.2
gunicorn==23.0.0
Werkzeug==3.1.6

# REST API & Security
flask-jwt-extended==4.7.1
flask-limiter==3.8.0
flask-cors==5.0.1

# 2FA Authentication
pyotp==2.9.0
qrcode[pil]==8.0

# Analytics
requests==2.32.3

# Testing
pytest==8.3.4
pytest-flask==1.3.0
